Introduce ApplicationDelegateController
